Combine existing DynamoDb implementation with new AmazonDynamoDb and expands tests
