+++ lib/Doctrine/KeyValueStore/Storage/DynamoDbStorage.php
@@ -21,8 +21,8 @@
 namespace Doctrine\KeyValueStore\Storage;
 
 use Aws\DynamoDb\DynamoDbClient;
-use Aws\DynamoDb\Exception\ResourceNotFoundException;
-use Aws\DynamoDb\Iterator\ItemIterator;
+use Aws\DynamoDb\Marshaler;
+use Doctrine\KeyValueStore\InvalidArgumentException;
 use Doctrine\KeyValueStore\NotFoundException;
 
 /**
@@ -33,18 +33,183 @@ use Doctrine\KeyValueStore\NotFoundException;
 class DynamoDbStorage implements Storage
 {
     /**
+     * The key that DynamoDb uses to indicate the name of the table.
+     */
+    const TABLE_NAME_KEY = 'TableName';
+
+    /**
+     * The key that DynamoDb uses to indicate whether or not to do a consistent read.
+     */
+    const CONSISTENT_READ_KEY = 'ConsistentRead';
+
+    /**
+     * The key that is used to refer to the DynamoDb table key.
+     */
+    const TABLE_KEY = 'Key';
+
+    /**
+     * The key that is used to refer to the marshaled item for DynamoDb table.
+     */
+    const TABLE_ITEM_KEY = 'Item';
+
+    /**
      * @var \Aws\DynamoDb\DynamoDbClient
      */
     protected $client;
 
     /**
-     * Constructor
-     *
-     * @param \Aws\DynamoDb\DynamoDbClient $client
+     * @var Marshaler
      */
-    public function __construct(DynamoDbClient $client)
-    {
+    private $marshaler;
+
+    /**
+     * @var string
+     */
+    private $defaultKeyName = 'Id';
+
+    /**
+     * A associative array where the key is the table name and the value is the name of the key.
+     *
+     * @var array
+     */
+    private $tableKeys = [];
+
+    /**
+     * @param DynamoDbClient $client    The client for connecting to AWS DynamoDB.
+     * @param Marshaler|null $marshaler (optional) Marshaller for converting data to/from DynamoDB format.
+     * @param string         $defaultKeyName   (optional) Default name to use for keys.
+     * @param array $tableKeys $tableKeys (optional) An associative array for keys representing table names and values
+     * representing key names for those tables.
+     */
+    public function __construct(
+      DynamoDbClient $client,
+      Marshaler $marshaler = null,
+      $defaultKeyName = null,
+      array $tableKeys = []
+    ) {
         $this->client = $client;
+        $this->marshaler = $marshaler ?: new Marshaler();
+
+        if ($defaultKeyName !== null) {
+            $this->setDefaultKeyName($defaultKeyName);
+        }
+
+        foreach ($tableKeys as $table => $keyName) {
+            $this->setKeyForTable($table, $keyName);
+        }
+    }
+
+    /**
+     * Validates a DynamoDB key name.
+     *
+     * @param $name mixed The name to validate.
+     *
+     * @throws InvalidArgumentException When the key name is invalid.
+     */
+    private function  validateKeyName($name)
+    {
+        if (!is_string($name)) {
+            throw InvalidArgumentException::invalidType('key', 'string', $name);
+        }
+
+        $len = strlen($name);
+        if ($len > 255 || $len < 1) {
+            throw InvalidArgumentException::invalidLength('name', 1, 255);
+        }
+    }
+
+    /**
+     * Validates a DynamoDB table name.
+     *
+     * @see http://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Limits.html
+     *
+     * @param $name string The table name to validate.
+     *
+     * @throws InvalidArgumentException When the name is invalid.
+     */
+    private function  validateTableName($name)
+    {
+        if (!is_string($name)) {
+            throw InvalidArgumentException::invalidType('key', 'string', $name);
+        }
+
+        if (!preg_match('/^[a-z0-9_.-]{3,255}$/i', $name)) {
+            throw InvalidArgumentException::invalidTableName($name);
+        }
+    }
+
+    /**
+     * Sets the default key name for storage tables.
+     *
+     * @param $name string The default name to use for the key.
+     *
+     * @throws InvalidArgumentException When the key name is invalid.
+     */
+    private function setDefaultKeyName($name)
+    {
+        $this->validateKeyName($name);
+        $this->defaultKeyName = $name;
+    }
+
+    /**
+     * Retrieves the default key name.
+     *
+     * @return string The default key name.
+     */
+    public function getDefaultKeyName()
+    {
+        return $this->defaultKeyName;
+    }
+
+    /**
+     * Sets a key name for a specific table.
+     *
+     * @param $table string The name of the table.
+     * @param $key string The name of the string.
+     *
+     * @throws InvalidArgumentException When the key or table name is invalid.
+     */
+    private function setKeyForTable($table, $key)
+    {
+        $this->validateTableName($table);
+        $this->validateKeyName($key);
+        $this->tableKeys[$table] = $key;
+    }
+
+    /**
+     * Retrieves a specific name for a key for a given table. The default is returned if this table does not have
+     * an actual override.
+     *
+     * @param string $tableName The name of the table.
+     *
+     * @return string
+     */
+    private function getKeyNameForTable($tableName)
+    {
+        return isset($this->tableKeys[$tableName]) ?
+          $this->tableKeys[$tableName] :
+          $this->defaultKeyName;
+    }
+
+    /**
+     * Prepares a key to be in a valid format for lookups for DynamoDB. If passing an array, that means that the key
+     * is the name of the key and the value is the actual value for the lookup.
+     *
+     * @param string $storageName
+     *
+     * @return array The key in DynamoDB format.
+     */
+    private function prepareKey($storageName, $key)
+    {
+        if (is_array($key)) {
+            $keyValue = reset($key);
+            $keyName = key($key);
+        } else {
+            $keyValue = $key;
+            $keyName = $this->getKeyNameForTable($storageName);
+        }
+
+        return $this->marshaler->marshalItem([$keyName => $keyValue]);
     }
 
     /**
@@ -76,14 +241,9 @@ class DynamoDbStorage implements Storage
      */
     public function insert($storageName, $key, array $data)
     {
-        $this->createTable($storageName);
-
-        $this->prepareData($key, $data);
-
-        $result = $this->client->putItem([
-            'TableName'              => $storageName,
-            'Item'                   => $this->client->formatAttributes($data),
-            'ReturnConsumedCapacity' => 'TOTAL',
+        $this->client->putItem([
+          self::TABLE_NAME_KEY => $storageName,
+          self::TABLE_ITEM_KEY => $this->prepareKey($storageName, $key) + $this->marshaler->marshalItem($this->prepareData($data)),
         ]);
     }
 
@@ -92,23 +252,9 @@ class DynamoDbStorage implements Storage
      */
     public function update($storageName, $key, array $data)
     {
-        $this->prepareData($key, $data);
-
-        unset($data['id']);
-
-        foreach ($data as $k => $v) {
-            $data[$k] = [
-                'Value' => $this->client->formatValue($v),
-            ];
-        }
-
-        $result = $this->client->updateItem([
-            'TableName' => $storageName,
-            'Key'       => [
-                'id' => ['S' => $key],
-            ],
-            'AttributeUpdates' => $data,
-        ]);
+        //We are using PUT so we just replace the original item, if the key
+        //does not exist, it will be created.
+        $this->insert($storageName, $key, $this->prepareData($data));
     }
 
     /**
@@ -116,11 +262,9 @@ class DynamoDbStorage implements Storage
      */
     public function delete($storageName, $key)
     {
-        $result = $this->client->deleteItem([
-            'TableName' => $storageName,
-            'Key'       => [
-                'id' => ['S' => $key],
-            ],
+        $this->client->deleteItem([
+          self::TABLE_NAME_KEY => $storageName,
+          self::TABLE_KEY      => $this->prepareKey($storageName, $key),
         ]);
     }
 
@@ -129,20 +273,19 @@ class DynamoDbStorage implements Storage
      */
     public function find($storageName, $key)
     {
-        $iterator = new ItemIterator($this->client->getScanIterator([
-            'TableName' => $storageName,
-            'Key'       => [
-                'Id' => ['S' => $key],
-            ],
-        ]));
+        $item = $this->client->getItem([
+          self::TABLE_NAME_KEY      => $storageName,
+          self::CONSISTENT_READ_KEY => true,
+          self::TABLE_KEY           => $this->prepareKey($storageName, $key),
+        ]);
 
-        $results = $iterator->toArray();
-
-        if (count($results)) {
-            return array_shift($results);
+        if (!$item) {
+            throw NotFoundException::notFoundByKey($key);
         }
 
-        throw new NotFoundException();
+        $item = $item->get(self::TABLE_ITEM_KEY);
+
+        return $this->marshaler->unmarshalItem($item);
     }
 
     /**
@@ -156,51 +299,24 @@ class DynamoDbStorage implements Storage
     }
 
     /**
-     * @param string $tableName
-     */
-    protected function createTable($tableName)
-    {
-        try {
-            $this->client->describeTable([
-                'TableName' => $tableName,
-            ]);
-        } catch (ResourceNotFoundException $e) {
-            $this->client->createTable([
-                'AttributeDefinitions' => [
-                    [
-                        'AttributeName' => 'id',
-                        'AttributeType' => 'S',
-                    ],
-                ],
-                'TableName' => $tableName,
-                'KeySchema' => [
-                    [
-                        'AttributeName' => 'id',
-                        'KeyType'       => 'HASH',
-                    ],
-                ],
-                'ProvisionedThroughput' => [
-                    'ReadCapacityUnits'  => 1,
-                    'WriteCapacityUnits' => 1,
-                ],
-            ]);
-        }
-    }
-
-    /**
+     * Prepare data by removing empty item attributes.
+     *
      * @param string $key
      * @param array  $data
+     *
+     * @return array
      */
-    protected function prepareData($key, &$data)
+    protected function prepareData($data)
     {
-        $data = array_merge($data, [
-            'id' => $key,
-        ]);
+        $callback = function ($value) {
+            return $value !== null && $value !== [] && $value !== '';
+        };
 
-        foreach ($data as $key => $value) {
-            if ($value === null || $value === [] || $value === '') {
-                unset($data[$key]);
+        foreach ($data as &$value) {
+            if (is_array($value)) {
+                $value = $this->prepareData($value);
             }
         }
+        return array_filter($data, $callback);
     }
 }
